[AMDGPU] Optimize masked transfer read in presence of fat raw buffers (#20604)
Adds a pattern that makes use of the out-of-bounds handling provided by
raw fat buffers in cases of all ones or all zero masks. This pattern
prevents elementwise selection code that otherwise gets generated when
we lower the masked transfer read to load + (elementwise) select. Note
that this pattern needs to run after bufferization but before
optimizations such as `OptimizeVectorTransferPass` run as they add
vector.extracts between the mask and the read making running such an
optimization harder.

---------

Signed-off-by: Nirvedh <nirvedh@gmail.com>