Reduce verbosity of HAL translation failure errors (#20614)
Some failure messages here are currently dumping entire functions,
making it difficult to see the actual error. This removes the IR from
the diagnostics.

To see the failing IR when debugging, you can pass
`--mlir-print-ir-after-failure` to `iree-compile`/`iree-opt`; this
approach also has the advantage of printing the IR in it's pretty form
when possible, instead of the generic form that error diagnostics use.