[build] Fix Bazel dependencies for shared library builds (#20596)
This brings in several fixes for dynamic library builds by listing the
dependencies explicitly. ATM there's no good way to regression-test
this, as the bazel config is only being checked statically, and the
CMake path seems to resolve the missing deps under the hood. That said,
keeping up with all the dependencies explicitly is a good practice from
a Bazel standpoint anyway.

---------

Signed-off-by: Artem Gindinson <gindinson@roofline.ai>