Integrate llvm/llvm-project@7b70fc7 (#20674)
Carries the same 4 reverts as
https://github.com/iree-org/iree/pull/20657:

- llvm/llvm-project#133231. This PR breaks fp_to_subbytes and
emulation_subbyte_types on llvm-cpu tests. iree-test-deps. tracker issue
in #20645.
- llvm/llvm-project#137122. StableHLO and Torch-mlir needs to update
their usage of GreedyRewriteConfig to use fluent API. i.e
enableRegionSimplification = VS setRegionSimplificationLevel
- llvm/llvm-project#135970. StableHLO has issue with VHLO_IntegerAttr
and APInt not being updated. StableHLO needs to be updated with that
PR's change for us to be able to integrate.
- llvm/llvm-project#121389. Torch-MLIR needs to be
updated with that PR's change for us to be able to integrate.

Signed-off-by: Ian Wood <ianwood2024@u.northwestern.edu>