[Encoding] Drop resolver interface implementation for SpecializedEncodingAttr (#20718)
The attributes were created for testing SpecializeEncoding pass when the
ideas were rough. The specialized_encoding is NOT a resolver, so it does
not implement the interface.

The revision moves the implementation to unspecialized encoding
resolver, and make corresponding changes. It returns itself in the
cloning methods. There are no new tests because the existing tests are
robust and already cover it. E.g.,


https://github.com/iree-org/iree/blob/360288f1a4e7e177a24ca290aabc083bd6e26b4f/compiler/src/iree/compiler/Dialect/Stream/Transforms/test/specialize_encodings.mlir#L153-L163

Signed-off-by: hanhanW <hanhan0912@gmail.com>