[NFC][Encoding] Move materializeEncodingValueFn to type converter (#20720)
This PR performs some refactoring as part of collapsing
MaterializeEncodingIntoPadding into MaterializeEncoding
(https://github.com/iree-org/iree/issues/20160) and moves the
`materializeEncodingValueFn` parameter from OpMaterializeEncodingPattern
to the MaterializeEncodingTypeConverter, together with the
`getInnerTileSizesOfr` implementation, for following reasons:
- The `typeConverter` and `materializeEncodingValueFn` objects are
typically passed around together, so now only `typeConverter` needs to
be passed around and this simplifies the code.
- As part of https://github.com/iree-org/iree/issues/20160, I am
creating a `getOffsetsSizesStrides` method in the
MaterializeEncodingTypeConverter and SerializableEncodingAttrInterface
and all encoding specific logic should be moved there. I would
specifically like to avoid the need for passing a
`MaterializeEncodingValueFn` to `getOffsetsSizesStrides`. This PR moves
us in that direction.

Signed-off-by: Jorn Tuyls <jorn.tuyls@gmail.com>