[Docs] Make op/attr/type summary styles consistent (#20726)
We have been publishing dialect docs for some time and the summary
fields across our dialects were inconsistently styled. This changes all
summaries to read more like a sentence (capitalized w/ punctuation) and
cleans up a few summaries/descriptions. Switches to use braces
everywhere as they make multi-line descriptions easier.

- flow.reshape/bitcast descriptions are improved
- iree_codegen.query_tile_sizes description is improved
- util.global.load description and summary are slightly less redundant

There is still a lot of room for improvement on the contents of the
docs, but this is a good first step.

---------

Co-authored-by: Han-Chung Wang <hanhan0912@gmail.com>